requirements.txt
pytorch-lightning==1.3.7
Pillow==8.2.0
tensorboard==2.4.1
matplotlib==3.4.1
numpy==1.20.2
nltk==3.6.2
